Emepzando schedsulerssfdsdskj

+++ sched.h
@@ -13,7 +13,7 @@
 #define player_A 0
 #define player_B 1
 
-unsigned short sched_proximo_indice();
+unsigned int sched_proximo_indice();
 
 
 #endif	/* !__SCHED_H__ */
